+++ test/jest/helpers/translationsProperties.js
@@ -1,5 +1,4 @@
 import componentsTranslations from '@folio/stripes-components/translations/stripes-components/en';
-import smartComponentsTranslations from '@folio/stripes-smart-components/translations/stripes-smart-components/en';
 import stripesCoreTranslations from '@folio/stripes-core/translations/stripes-core/en';
 import innreachTranslations from '../../../translations/ui-inn-reach/en';
 
@@ -7,10 +6,6 @@ export const translationsProperties = [
   {
     prefix: 'stripes-components',
     translations: componentsTranslations,
-  },
-  {
-    prefix: 'stripes-smart-components',
-    translations: smartComponentsTranslations,
   },
   {
     prefix: 'stripes-core',