+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for ui-inn-reach
 
 ## (in progress)
+* Bump stripes to 8.0.0 for Orchid/2023-R1. Refs. UIINREACH-215
 
 ## [2.0.3] (https://github.com/folio-org/ui-inn-reach/tree/v2.0.3) (2022-12-22)
 [Full Changelog](https://github.com/folio-org/ui-inn-reach/compare/v2.0.2...v2.0.3)